requirements.txt
Flask==2.3.2
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry.instrumentation.flask
